requirements.txt
certifi==2025.1.31
cftime==1.6.4.post1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.7
kiwisolver==1.4.8
matplotlib==3.10.0
netCDF4==1.7.2
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dateutil==2.9.0.post0
shapely==2.0.7
shiboken6==6.9.1
six==1.17.0
